Let winmanager do copy of paint-buffer to VESA-mem (fixes #6)
This solves the problem of the paint-order which existed before
because the winmanager just sent events to the applications and asked
them to do the copy. Of course, the order depends on the scheduling in
this case, so that it might be wrong.

Now, the winmanager creates a shared memory region for each window and
the application maps that in as well, so that the application can
directly paint in this region and ask winmanager to copy a specified
portion of that to the VESA memory. This does of course mean that we
have to recreate the shared memory region on every resize operation, but
since we have the preview-rectangle-concept, it doesn't really hurt.