Introduced dirty-flag and handled lost updates.
That is, every UIElement has a dirty flag and only if it is set,
repaint() has an effect. This relieves the user from thinking
about whether it is called too often and thus prevents useless
repaints. Additionally, it makes it easier for controls to track
whether something has changed (like Editable).

To prevent repaints during updates, we simply disabling them
during that time and if someone tries, note it and do a repaint
as soon as we've got the response from the update.