Improved signal-handling.
Now we don't check at every kernel-leave for our own and the
signals of others. Instead, we have a quick for our own and
ignore others. Every added signal makes the corresponding thread
ready. For the interrupt-signals, we switch directly to the next
thread to reduce the latency.
Additionally, the locku function of MMIX has been improved to
use yield if the lock is held. We don't have multiple CPUs there,
so waiting doesn't make any sense. This turned out to be a
problem after changing the signal-implementation. For whatever
reason. But it's definitely better and fixes the problem.