Generalized lists and used DList for threads and signals.
Now we have SList, DList, ISList and IDList. For better
thread-destruction-performance we use DList now for the list
of threads and the list of signal-threads. This works by
putting an DListItem-inherited class with a pointer to the thread
into the thread-class (two of them). Therefore, we can remove
items in O(1) from these lists and can access the thread from
a list-item.