Use hasSignalQuick() where possible.
The only place where we need to make sure that no signal
arrives until we've put ourself to sleep is in the scheduler.
Because otherwise we might sleep for ever. At all other places
it doesn't really matter whether we handle it now or later in
the case where we get a signal exactly in that moment. So,
the inlineable version without lock is better there.