Don't open/close a fd for every request.
Now we open the file for the driver when the client connects
and store it in the channel. The driver receives the fd from
getwork which allows him to communicate with the client and
also to identify him (the fd won't change). This avoids a lot
of code in the IPC path, especially since quite a few locks
were required.
The close is a bit tricky, but I think I've got it working
correctly. Additionally, I had to change the "used-concept"
for channels. Since we don't have a clear ending of a request-
handling, we can't mark the channel unused again. Thus, I
introduced a flag for getwork that enables the used-marking
and added a fcntl command to reset it. This is only necessary
for parallel request handling where the requests have to be
quite compute intensive anyway to compensate (at least) the
thread-switch, so that it doesn't hurt. And this way, it
doesn't affect the default, non-parallel performance.
overhead
Of course, the downside is that we might need more
file-descriptors now. Perhaps we should increase the fd-array
on demand now to allow more clients.