Made local and global semaphores available for userspace.
That is, now we have process-local semaphores for synchronization.
Additionally, we have global semaphores, which are implemented as
VFS nodes and can thus be shared, have permissions, can be viewed
in the FS and so on.
locku uses process-local semaphores as a backend now. The rw-lock
doesn't work for that, as I've noticed now. And not at all in the
way I've used it. Strange that this never triggered in the last
weeks...