Added a tool that adds shared library symbols to gdb.
Since our dynamic loader is deterministic, we can calculate
beforehand what the addresses of the shared libraries of an
application will be. Thus, I wrote a shellscript that does
that and call it from b to add the symbols from all loaded
shared libraries of the application that should be debugged
to gdb.