Improved timer-handling on x86 and interrupts in general.
Now, we use the LAPIC for timer interrupts, so that we actually
have preemption on APs ;) Additionally, the organization of the
timer-stuff has been improved.
The PIT is still supported in case there is no LAPIC or if the
kernel-parameter "forcepit" has been provided.
The interrupt-code is a bit less architecture-specific now and
we have /system/irqs that tells us how many times each IRQ
was seen.