Added a app to change the screenmode including infrastructure.
That is, the winmng has a command now to set the screenmode. It
will ask all windows to do a reset afterwards, i.e. get the new
mode, recreate the window-buffer and repaint all. To do that,
we had to get rid of the subclasses of Graphics for the different
color depths. Otherwise we would have to give each uielement a
new instance of it.
For the settings-application, radiobuttons have been added
including paint-support, i.e. draw/fill-circle.