Updated all cross-compilers to binutils 2.23.2 and gcc 4.8.2.
This was actually quite difficult because I had to update the
eco32-backend for gcc from 4.4.3 to 4.8.2, which was not really
straight forward. But it was necessary sooner or later because
the C++11 support in 4.4.3 is far from complete, which caused
more and more headaches recently.
There were also quite a few changes from 4.6.3 to 4.8.2
regarding libgcc, dynamic linking and stuff, which was not easy
to figure out, but know its working correctly, as it seems :)