Improved priority adjustment scheme.
Now we do that based on the blocking time instead of the CPU
time, which behaves much better if someone creates a large
amount of CPU intensive threads (previously, the system was
basically dead in this case). For details, see comments.
Additionally, processes receive the minimum priority of all their
threads and inherit this priority to created threads and child
processes (bad guys are likely to spawn more bad guys ;)).