Fixed issue in reboot/shutdown.
We can't terminate the disk and fs process, for example, because
we need them for demand loading and stuff. That is, if init e.g.
touches code that it hasn't touched before during reboot/shutdown,
we need to do demand loading which only works if these processes
are still alive.
Therefore I decided to keep all processes started from boot-modules
always alive (and flush the fs cache via sync()).
For that matter, I had to add files to /system/mbmods for eco32
and mmix because they were missing so far.