Use reference counting for threads and processes.
We have to be sure that the objects stay valid during their usage.
Previously, there were some corner-cases where this wasn't true.
For example, VFS-info accessed the process-/thread-properties
without locking or similar. Additionally, the access to the
process- and thread-array has to be atomic. That is, the procedure
of checking whether a process/thread exists and the acquiring of
its mutex has to be atomic to make sure that nobody gets/uses an
invalid pointer.