Added an easy way to share a buffer with a device.
This way, if high performance is required, data can be exchanged
with a device over shared memory, which prevents one copy. The
nice properties are that it doesn't increase the complexity of
the kernel-code and requires nearly no changes to applications.
Because the only thing an application needs to do is call
sharebuf() with the fd to the device and use the created buffer
for read/write. Everything else works automatically.
It's also quite easy for drivers to support this. They only need
to react on MSG_DEV_SHFILE by joining the received shared memory
and handle MSG_DEV_{READ,WRITE} differently depending on whether
the shared memory should be used (the kernel tells them) or not.
Note also that applications do not have to react on whether a
device supports it or not. If not, it works in the same way for
the application with the only difference that the created
buffer is not shared but used as ordinary source/destination
for copying in the kernel.
So, if high performance is required one can pay the one-time
overhead for sharebuf() to transfer data faster afterwards via
read or write.
A disadvantage is that one can only share one buffer per channel,
which is also problematic for channel-sharing. But this does not
occur for performance-critical stuff typically, so that it isn't
that bad.