Made ext2 and iso9660 standalone.
This is the first step in the fs-redesign. Now, ext2 and iso9660
are standalone programs (linked against libfs), which is much
more flexible and scalable than the single-fs-solution.
Currently, mounting is basically disabled because of the
separation. But in the next step we'll integrate the mountpoint-
management into the kernel, so that it works again.