Moved mounting into kernel.
That is, now we have a mountspace per process which is inherited
to child processes. We have the system calls clonems(), getmsid()
and joinms(). The first one detaches the mountspace, i.e. clones
it so that the process has its own copy and can manipulate it
independently. getmsid() can be used to e.g. pass the mountspace
id to another process which can then use joinms() to join that
same ms. This is intended for e.g. letting login create a
mountspace per user by checking whether it already exists,
creating it when necessary or otherwise joining the existing
one.
The reason for having the mounts in kernel is that the kernel needs
to access it as well for e.g. executable and library loading. So, it
needs a complete and working mountspace and moving that to a
userspace mount service or something, would complicate it a lot
since the virtual filesystems have to be mounted there too and
of course it would introduce quite a bit of overhead.
The virtual filesystems are mounted by the kernel at the beginning.
The same is true for the root-fs. All of them can be unmounted if
desired, although some things won't work anymore afterwards,
obviously.
Unfortunatly, one can't remount the virtual filesystems or have
multiple /dev-fs or something. The reason is that they are
currently handled in a special way in the kernel. It would be
possible to have in-kernel services for that, i.e. use the same
message passing as to userspace-filesystems. I'm not sure yet
whether the overhead is worth the increase in flexibility.