Keep used and exclusive files in a linked list.
This way, we don't have to look through all files up to the
biggest number, but in most cases only the exclusive ones, which
should be basically none. If we want to use the file exclusively,
we have to search through the used ones, obviously.
This does also mean that we don't use the same file anymore if
the one process opens it multiple times. But IMO this was really
unintuitive anyway since they shared the read/write-position
which the user-application of course needed to know to not do
something stupid. But what would have been the point of opening
a file twice then?