virtmem: added flags MAP_LOCKED and MAP_NOSWAP.
With the first one, we can lock regions into memory. The second
one causes the kernel, if used together with MAP_POPULATE, not
to start swapping if there is not enough memory but just fail.
Additionally, some issues have been fixed.
Note also that we're now demand-loading all regions, i.e. also
regions that aren't loaded from a file. These are simply zeroed
as we've already done that for the end of file-based regions.