Fixed the overpaint-bug when switching between UIs.
This is mostly done by using F_DISMSG before the change. This way,
we ensure that no messages are still in flight for the old client
and thus no updates will be made for the old client when we've
already switched to the new one.
Additionally, we are a bit more careful with setting the active
client during that time.