Removed floating point stuff from kernel.
We can't use it there that easily because we would have to
save the user-state first and restore it afterwards. So, better
don't use it at all.
To ensure that we don't use it by accident in the future, I added
the -msoft-float option which results in a linker error if we
have floating point arithmetic.
For that reason, we had to move some stuff from basic into the
c library and some tests out of the kernel test into libctest.