Redesigned synchronization primitives for the userspace.
The kernel provides process-local semaphores that are inherited
to child-processes in a shared way. They don't survive exec's but
one can still use them for synchronization with non-exec'ed child-
processes. Mostly it is used within one process, of course.

Additionally, the kernel has global semaphores that are attached
to the filesystem. That is, one can synchronize on any file in
the tree by opening it and calling fsemup/fsemdown on that file-
descriptor. This does not affect reading, writing etc. of that
file. We just use the path of the file as the global name, in a
sense. This way, in many cases it will be simpler to come up
with a reasonable and unique name (which does already exist
hopefully). It is created on demand currently. It might be
imaginable to have an init-fcntl for it to assign the initial
value for the semaphore and have a fixed point of initialization.
Currently, we don't need it.

The userspace has an additional abstraction based on process-
local semaphores and atomic operations for the non-contention-
case.

The reader-writer-lock in the kernel has been removed entirely
and in general the synchronization stuff is a lot cleaner now.
We'll implement the reader-writer-lock in userspace in a similar
way as the usem* primitives.