Removed wait() and notify() syscalls.
They complicate the kernel and are not really necessary. wait()
was typically used for cases where e.g. getwork() was called
non-blocking for doing some finishing tasks when there is no
more work currently. Another and arguably better solution is to
only call getwork() non-blocking if there is something to finish
and otherwise call it blocking.
The shell did also use that because fgetc() ignores signals and
wait() didn't. I think we should stick to letting fgetc() and
others ignore signals (stopping a stdio operation somewhere
in the middle because of a signal doesn't really work). So,
instead the shell uses read now and manages a small buffer
on its own.
Other uses of wait() and notify() might have been
synchronization, but semaphores are better for this case
because you can't loose an "event".
Additionally, I organized the syscalls in a way that we can
add/remove syscalls more easily.