Use a semaphore for reading from devices.
That is, the old set-readable concept is gone and we use a
semaphore that is up'ed by the driver via fcntl() and down'ed
by the kernel if a client wants to read something. This is
much easier to implement and arguably cleaner.
This is used by all character devices (atm only vterm). Other
devices always have data to read, so that this is not necessary.
To be able to implement that, I had to change the way references
of VFSNodes are counted as well. Now parent nodes have a
reference by all their childs. This allows us to assume that
our parent does always exist (although maybe not reachable
anymore), which simplifies things and especially doesn't require
us to grab a lock at some critical places.