Translate abc://def to /dev/abc/def.
This way, we can pass an URL to applications which would be
translated to a device path + arguments. So, in other words,
later we can e.g. provide an http driver to read files from http
and process it with any command line tool we have:
cat http://www.google.de | grep foo | wc -l
In this case, we would do an open to /dev/http with
www.google.de as the argument, so that http can open the
connection etc..