Removed user-semaphore from stdio.
It's a bit odd, but it wasn't used at all. At some point, we
should probably make stdio thread-safe, but I'm not sure if
we need a user-semaphore per FILE object. This would make
snprintf quite expensive, for example, because we need to
create and destoy a semaphore for it.