Introduced request-ids.
Now every message has a request-id and responses have the same
message-id as the request in order to pair request and response
easily. Fortunately, this is transparent for the applications.
The IPC API takes care of that.
Thus, we don't need to store the thread of the last client we
received a message from and we also don't need the channel-used
concept anymore. I.e. drivers are not forced anymore to reply
to received requests in the same order, which is especially
important for networking.