Removed read-semaphore and added cancel().
That is, now character-devices don't use the read-semaphore
anymore to only go to the driver if there actually is something
to read. This was too specific and also error-prone.
Instead we go directly to the driver and the driver queues the
requests correspondingly, which isn't that hard if we offer a
appropriate abstraction.
Of course we would like to interrupt long running read- or write
requests via signal. Thus, I introduced the cancel() syscall and
MSG_DEV_CANCEL. If a driver chooses to support MSG_DEV_CANCEL,
the kernel expects it to be able to handle interruptions of read-
and write. So, if the client gets a signal during the receive in
read, for example, we send MSG_DEV_CANCEL to the driver. The
driver is expected to reply with 1 if the response has already
been sent (bad timing), 0 if he has now canceled the request or
an error-code if there is something wrong. If the kernel receives
1, it will repeat the receive with signals disabled.
In order to allow that for other operations besides read and write,
we have the cancel() syscall. For example, the recvfrom and sendto
operation on a socket are also interruptable by signal. In this
case the userspace has to use cancel() to cancel the request.
All the other kernel-supported operations (open, sharefile, ...)
are not interruptable since it is expected that they deliver an
immediate response. Although I'm not entirely sure if we want to
have open() interruptable on some occasions...