Added creatsibl.
This syscall allows a client to create a new "sibling" channel
from an existing one. So, instead of creating a new channel via
open(), one can use creatsibl to create a channel with the driver
knowing that it belongs to an existing channel. This can be used
for creating a separate event-channel that is associated with a
control-channel and also to implement accept() for sockets.
It can be optionally supported by devices and is cancelable, like
read and write.