Don't terminate the process on segfaults.
That is, ensure that memory accesses that might fault in kernel
always use a known instruction so that we can skip this one
and continue executing to e.g. deliver a SIGSEGV to the
process. Additionally, we don't need to note down the resources
we have acquired before performing accesses that might fault,
because we'll always continue and finish the execution of the
kernel-entry (until the SIGSEGV is delivered).