Moved environment-variables to userland.
Now that we have the new UserAccess-concept, it is more
complicated than before. Additionally, if we do the
absolutize-path in userland, we don't need any env-vars in
kernel anymore. Of course, we might do additional work by
first absolutizing the path in userland and then copying
it again in kernel, but because the UserAccess stuff costs
a bit, too, we are actually not worse.