Added bindto() syscall.
This allows it to bind a device or channel to a thread. That is,
we store a thread-id for every channel and device. The thread-id
of the device is the default one for created channels. When one
calls getwork() one gets only messages from channels that belong
to the current thread.
This way, we can make use of multiple cores in an easy way by
e.g. assigning a channel on creation to a specific thread. Or
we could create a new thread specifically for a created channel
to logically separate the handling of channels.