Added SIG_CANCEL.
Now that we know the handling thread of a channel, we can
notify it via signal on cancel(), too. This allows drivers
to do a blocking syscall in the implementation of e.g. the
read-request and get interrupted if the client wants to
cancel (maybe because he got SIG_INTRPT).
I decided to make that orthogonal to the cancel-message.
That is, one can register a handler for SIG_CANCEL and
not support the cancel-message or the other way around
or support both. I don't know if there is a usecase for
getting both, but this seems to be the most flexible
solution.