Forward the size-operation on channels to the driver.
The reasoning behind that is the following: The stat() operation
does always return the information for the given path. When
using open() and fstat() on the other hand, open() may have the
side-effect of e.g. creating a channel and associating the fd
with that channel. Thus, fstat() returns information about the
open file, which might be something different than the path that
has been opened.
So, whenever one wants to have information exactly about a
specific path, one should use stat(). If one wants to have
information about an open file, one should use fstat().
This allows us to e.g. get the size of an HTTP file to load
when using open() and fstat() and still retrieve information
about /dev/http with stat().
Note also that I decided (for now) to only forward to size
gathering to the driver. The other infos are still set by the
kernel from on the VFS node. Because most of these infos are
completely meaningless for open files, like uid, permissions,
inode number etc.. Perhaps it makes sense to add blockSize and
the dates at some point, because these may have a meaning for
some kind of files (like the modified date of a website).