Made build-system and directory organization ready for x86_64.
That is, now we have a target and a target-type to support
code-sharing of architectures that are similar. We also use that
to share the drivers code of eco32 and mmix.
Additionally, I replaced the usage of #ifdef with #if defined,
and we define __TARGET__ and __TGTTYPE__ now instead of using
the builtin defines.