Clean up startup.
Now we don't use the gdt-trick anymore, but setup paging right
away so that we don't need to deal with different "environments"
in the C++ code.
Additionally, we don't use the external trampoline anymore. Now
it's integrated into the kernel.S, so that we don't need nasm
anymore and can use common code easier.
I've also moved the memory-layout constants to mem/layout.h,
which enables us to include it in assembly files.