Redesigned paging for i586.
We don't put the page-directory in itself anymore, but walk
through the page-tables for all PTE accesses. This works because
we allocate all page-tables in the lower memory, which is
directly mapped, so that we can easily access all page-tables.
The new approach has the advantage that we don't need to to map
and unmap other page-directories in ourself all the time and
we probably be able to share a large portion of the code with
x86_64. Additionally, we only need a lock for the temp-page
now (per page-dir because we only have one temp-page now, not
a temp-map-area with several page tables). Everything else is
fine without because we already do locking in the modules
that call us.