Changed the interface to paging.
Instead of passing an array of frame-numbers to map(), we have
an allocator class now. This is not only used for map(), but
also for unmap() and is responsible for allocating and freeing
pages and pagetables. We have different subclasses that
implement the different cases (user, kernel, kernel-stack,
no-alloc, range).
This gives us much more flexibility and simplifies it for some
usecases.