Moved x86-paging to a general pagetables module.
We use it for eco32 now as well, which prevents a lot
of code duplication. We don't use it for mmix atm, because
there are some differences. I'm not sure whether one could
consolidate the code, but for now we keep it that way.
Additionally, I cleaned up the paging-related header files.