ftpfs: added support for rename.
To make that work, I had to introduce the rename operation
for filesystems, which in turn meant to introduce the rename
syscall. The problem is that on UNIX filesystems, the operations
link+unlink are sufficient to implement rename. But the FTP
protocol doesn't support a link operation, but only rename, i.e.
the higher level operation, build with link+unlink. Since it
seems to be impossible to guess from consecutive calls of link+
unlink that it's actually a rename in an reasonable way, I chose
to add the higher-level operation.