tcpip: ensure that we don't resend data if we don't want to.
We only want to resend the data after a timeout. When we've got
a packet, we might want to send additional data, but not the same
again. Note that the problem did only occur if we got a "weird"
ACK. For example, if we got an ACK for the stuff before the data
we've just sent, then we would've sent the data again, without
reason.