ftpfs: use ABOR if we want to stop reading.
Sending a RST over the data-connection does not suffice. We need
to tell the server over the control-connection that we want to
stop.
It's a bit tricky to do that because if the transfer is still in
progress, we'll get just the response for ABOR. If it's already
finished, we get the response for the transfer, too. That means,
we have to know when a transfer is complete. To do so, I use the
size of the file from the directory listing.