Moved rtc into the kernel.
There is no point in having these 30 lines in userspace. The
kernel needs date+time as well for e.g. the timestamps of
the VFS-nodes. Additionally, getting the time is a performance
critical operation for e.g. filesystems. Therefore, I think the
best way is to integrate that into the kernel.
That is, now we have a gettimeofday() syscall, have proper
timestamps for the VFS-nodes and don't need the performance
optimized time() version in ext2 anymore.
Since the kernel needs mktime now, too, I've moved the time-
computation stuff into basic.