Moved TLS support from the kernel to userland.
We'll do that in userland by putting a pointer on the top
of the stack now. Since the stack has a power-of-2 maximum size,
we know the stack top and can thus use that to access TLS
very fast.
The disadvantage is that we can't use the compiler support for
TLS anymore (__thread). The advantages are that it removes
quite some special cases from the kernel, works for all
architectures, with and without dynamic linking and removes
some code from the critical context-switch path.
Additionally, I've put errno on the top of the stack. Thus, it's
thread-local now, too. Note that it is separate from TLS because
we need errno to support TLS.