Integrated mountspaces into the VFS.
That is, every process has a link to a MS-VFS-node as its
mountspace. For every operation (mount, umount, clonems, joinms)
the user specifies the mountspace in form of a file-descriptor
pointing at a MS-VFS-node. Thereby, we can reuse the permission
system from the VFS (chmod, chown, stat, ...) and thus control
who can access/change what mountspace.
Every mountspace is living in /sys/ms as a file, whose name is chosen
when using clonems. Additionally, each process has a file in its VFS
folder (/sys/proc/<pid>/ms) which links to the mountspace file.