Added own iostream implementation to libesc.
It has several advantages:
- like standard iostreams, it is typesafe, in contrast to printf
- unlike standard iostream, it provides a simple and concise
way to do formatting
- additionally, it does not have an internal formatting state,
where the user has to know which formatting option is applied
permanently and which only for the next object to format.
- since fstream uses a FILE object internally, one can easily
mix usages of c-functions working with the FILE object and
the fstream. One can also construct an fstream from a FILE
object or file-descriptor.
- furthermore, its implementation is much simpler.
Thus, it is much easier to use, so that I replaced all usages
of standard iostreams with my own ones. Additionally, I added
unittests for it.