Removed stat system call in favor of fstat.
Overall, I'm trying to reduce the number of syscalls that
deal with paths to simplify the kernel (and partially, drivers
as well). This will decrease the performance of these operations
slightly, but none of them are really performance-critical (stat,
link, unlink, mkdir, rmdir, rename, chown, chmod, utimes).
The first one is stat, which has been removed so that only fstat
is supported by the kernel. The C library provides stat, though, and
internally does an open, fstat, close.

Stat is a special case, though, because open on a device creates
a channel by default. If we do an fstat on the resulting fd, we
of course get the information about the channel, not the device.
Furthermore, creating a channel might not even work (permissions,
missing path behind the device, ...). Thus, stat uses a special
open flag now to now create a channel, but get a fd for the
device itself. It can't do much with it, though, since it does
not support or have permission for read, write, bind, getwork
and so on.