ftpfs: create device after connect.
Because until we are connected, nobody can use the device
anyway. And currently, it caused the problem that mount was
waiting for ever for the shell without being able to receive
signals (in open). There is still something wrong, because
actually the kernel should notice that the communication
partner is gone and deliver an error back to mount.