kernel: use events for semaphores, not the blockedlist.
The problem is that we want (in some cases) allow to
receive signals while waiting on a semaphore. If that
happens, the scheduler is notified to unblock the thread
and has thus to know how to do that. Since it doesn't know
about the blockedlist, it can't dequeue it there.
Thus, to avoid these problems, just use the event system
we already have, which simplifies it and works well with
receiving signals while waiting.