Improved file protocol.
We want to be able to reply with an error code for any request
without knowing the protocol. For this reason, we need to have
an error code as the first value in each reply and need to agree
on the type.
This is the first commit to allow that, which changes the file
protocol correspondingly.