Moved syscall tracing from kernel to userland.
Previously, we had a really simple syscall tracing facility in
the kernel, which has now been moved to userland. Currently, it
is implemented in the libc, which checks if tracing is enabled
before and after each syscall. In this way, it is still simple
(no need for something like ptrace) and works for statically and
dynamically linked programs (unlike an LD_PRELOAD-like approach).

One can either trace all syscalls of all processes by specifying
the kernel parameter logsysc or use strace to trace syscalls of
a process and its childs. Both is handled by an environment
variable, specifying the file descriptor to write the trace to.

To print syscalls that caused a context-switch in a nice way and
support multiple threads/CPUs, a synchronization file is used
that hosts a generation counter. With this, context-switches can
be detected and reflected in the written trace.

It is a bit more complicated in userland, because we need to take
care what functions we call in the tracing. For example, we can't
use the heap, because the heap can cause a syscall, which should
be traced. But the advantage is that we can add all kinds of
conveniences like decoding parameters, showing names of flags
and so on without increasing the size and complexity of the
kernel.

Currently, this works on x86 only. But adding support for eco32
and mmix should be easy.