Changed user/group handling.
Previously, we had the more or less usual /etc/{passwd,users,
groups} and used the setuid-bit for the users and groups
utilities in order to allow all to e.g. list users and change
their own password, but not add/remove users etc.

Since setuid is a really bad thing and unnecessary in this case,
I've changed that now. Now, we use the permission system of
the filesystem to control who can do what. For that to work, we
have dedicated folders and files for the individual objects and
attributes. E.g. /usr/<user>/id is owned by root and readable by
everybody, but /usr/<user>/passwd is owned by <user> and only
readable/writable by him.

This also simplifies the usergroup library significantly, since
1. both users and groups can be accessed and manipulated in
the same way and thus with the same code and 2. no complicated
parsing of file contents is required anymore. Also, most of the
actions in the users/groups utilities are simple file
operations now.