Improved mount concept.
I now added the missing restriction, that the directory where
a filesystem is mounted needs to be writable for the user. Also,
if it's a sticky directory, it needs to be owned by the user.

Further, mount is no longer reusing existing filesystems, but
is always starting a new one. For the purpose of reusing
existing filesystems, we now have the application bind.