Access control semaphore IRQs.
That is, the semcrtirq syscall now expects a file descriptor for
the interrupt to attach to. And at boot, the kernel creates a
file for each interrupt. For attaching to an IRQ, the process
needs exec permission.

At the moment, only root can do that. But later, init could set
the permissions correspondingly, to give e.g. the disk driver
only access to the disk IRQ and so on.