uimng: fixed mean bug in child creation.
Forking with multiple threads is really bad. Because the other
threads might be in an arbitrary state at the point where we make
a snapshot of the memory. Thus, the program state might be
inconsistent, we might have taken locks and so on.

To avoid these problems, uimng does now take care to only do simple
system calls in the child before the exec.