Make use of delegate in the UI.
That is, we do no longer pass paths around and let both parties
open and mmap it, but let the client create a temporary file and
delegate it to the driver, which can then map it. Currently, the
temp file is shortly visible between create and unlink. This
could be improved later by offering a system call that directly
creates an invisible file.

Overall, this simplifies it a lot and prevents name clashes and
their bad workarounds.