gui: fixed repaint bug.
We have to make sure that _minoff is >= _off in Graphics. Because
we clip the size of the paint rectangle as well, so that we have to
clip the position, too. Otherwise, repaint requests to controls in
a scrollpane that are scrolled by more than the size of the scrollbar
are ignored.