kernel: moved kernel threads out of init process.
Now we use the first process (0) for idling, process 1 for init,
process 2 for the terminator and, optionally, process 3 for the
swapper. This makes it less confusing for the user since we do
no longer mix various different things (people might have
wondered why init consumed nearly all CPU time, for example).
It also allows us to assign proper names to the processes.