Start network subsystem at boot and as non-root.
We have a network startup service now, that starts the network
drivers found on the PCI bus and the loopback device. It registers
the links at tcpip by itself and also restarts drivers if they die.

This does also allow us to run the network stuff with the network user
instead of root.