Make file access depend on mount-file, not mounted fs.
That is, if we open a file for writing for example, it is no longer
required to have write access to the mounted fs device
(/dev/ext2-hda1), but the file (OpenFile instance) which was opened
for the fs device needs write access.

That makes it more consistent with the rest of the permission system
and gives us more flexibility.