Replaced vtctrl with uictrl.
Previously, vtctrl talked to the vterm (hence the name) to do
mode listing and setting. This was unfortunate in the GUI because
the GUI terminal (part of guishell) has nothing to do with that
and thus didn't support it.

uictrl talks to the UI (either managed by vterm or winmng) to do these
operations. Additionally, it allows to list all UIs and start new ones.