libc: workaround for pthread with static linking.
When we link statically, we link against the gcc_eh, which references
pthread functions. These are implemented in the libc, but for some
reason, the linker does not find them and thus leaves the symbols
undefined in the executable. We work around that for now by simply
using one of the pthread functions in the libc.