Added sandbox application.
After months of preparation, we finally have everything in place
to do (basic) sandboxing :) The sandboxing mechanism of Escape
allows to run an application with the own user, but restrict its
permissions by leaving groups and reducing the permissions of
subtrees in the filesystem. The latter is done by remounting
directories in the mountspace with reduced permissions.