kernel: improved VFSFile implementation.
Previously, we allocated the memory on the kernel heap, which has
obvious limitations. Now we store it hierarchically in pages and
allocate the memory on demand.

This is especially good for large temporary files like the buffer
cache, which is atm allocated in the kernel-provided tmp directory.
That we could previously allocate it there without getting an error
because of the filesize was actually a bug :)