libc: use mmap in heap, not chgsize.
The chgsize syscall does not scale well with the region size. Since
we have mmap as well and the heap doesn't care whether it uses
contiguous memory or not, better use mmap. We increase the allocated
sizes exponentially until a certain size to support both simple programs
and programs that heavily use the heap well.