Improved shutdown/reboot.
I have removed the I-am-alive message, which was basically useless,
because everybody needs to be able to send it (think: fs mounted by
unprivileged user) and we have no idea how we should postpone the
shutdown/reboot. In the end, we either need to trust them or we have
a hard timeout. I chose the second option, in which case the whole thing
is unnecessary. Just SIGTERM all first and SIGKILL them a few seconds
later.

Furthermore, I've introduced a power group for the reboot/shutdown
device, so that we can control who is able to do that.